refactor(agent): move API config methods from ModelConfig to ModelHandler
* Move get_api_key and get_base_url to ModelHandler class for better separation of concerns
* Update all model handlers to use new method locations
* Remove os import from model_config.py
